clean up diff on tests
